forpython/for-beerbot/requirements.txt
python-telegram-bot
click
